на всякий

+++ src/main/java/com/example/SocialNetwork/api/NotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.SocialNetwork.api;
 
 public class NotFoundException extends RuntimeException {
-    public <T> NotFoundException(Class<T> clazz, int id) {
+    public <T> NotFoundException(Class<T> clazz, Long id) {
         super(String.format("%s with id %s is not found", clazz.getSimpleName(), id));
     }
 }
